fixed last bug
i've finally fixed the last critical bug of this project, the bug was a data race which would result in a dead lock on the lcd forcing freeRTOS to restart.

it was fixed simply by protecting the concurrent lcd calls with a mutex, since it's just display and it doesn't need much speed it's a seamless solution.

i also added conditions to only change values that have actually changed